WIP: Update tests to run with ROS2 Kilted
**Warning** The ROS2 tests will either crash due to a double
free error or I have to explicitly allow a memory leak by not deleting
the optimizer pointer. This is happening specifically with ROS 2 Kilted
on an AMD system and additional testing is needed to see if this also
affects ROS 2 Iron.

Other Changes:
- Update C++ standard for newer versions of gtest in Ubuntu 24.04
- ROS 2 Kilted changed the MPPI Controller API slightly so all the ROS 2
executables now use an environment flag to check what ROS2 distribution
exists and add a compilation flag to switch between the two APIs.
- Add compilation warning suppression flags to CUDA builds as Eigen
likes to write CUDA code that produces warnings when compiled with nvcc.
- Update CMake to no longer use the `ament_target_dependencies()`
command as it is deprecated.